log-consumer-system/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
redis[hiredis]==5.2.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dateutil==2.9.0
jinja2==3.1.5
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
fakeredis[lua]==2.26.2
structlog==24.4.0
pyyaml==6.0.2
